size_t -> unsigned
git-svn-id: file:///Users/gsalam/svn/fastjet/trunk/fastjet-release@214 9ab96302-e443-4389-8a14-f726967065a6

+++ areas/FjClusterSequenceWithArea.hh
@@ -198,7 +198,7 @@ template<class L> void FjClusterSequenceWithArea::_initialise
 
   if (writeout_combinations) {
     cout << "# Printing particles including ghosts\n";
-    for (size_t j = 0; j < _jets.size(); j++) {
+    for (unsigned j = 0; j < _jets.size(); j++) {
       printf("%5u %20.13f %20.13f %20.13e\n",
 	       j,_jets[j].rap(),_jets[j].phi(),_jets[j].kt2());
     }
